# apiDoc.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: myDeben
  description: "API Documentation"
  contact:
    email: "youssuf.radi@deemalab.com"
host: "localhost:3000"
basePath: "/api"
...
schemes:
- "http"
paths:
  /auth/user/signin:
    post:
      tags:
      - "Auth"
      summary: "User Sign In"
      description: "Returns a token for user to be used in further identification"
      operationId: "userSignIn"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "2 Scenarios are allowed:
                      - User email and password for sign in
                      - service and id to sign in with"
        required: true
        schema:
          $ref: "#/definitions/SignInService"
      responses:
        200:
          description: "Successfully Logged In"
          type: "object"
          properties:
            detail:
              type: "string" 
            auth:
              type: "boolean"
            token:
              type: "string"
        400:
          description: "Bad request or Required fields not stated"
        401:
          description: "Wrong Username or Password used"
        404:
          description: "Service not found"
  /auth/user/signup:
    post:
      tags:
      - "Auth"
      summary: "User Sign Up"
      description: "Creates a new user in the database"
      operationId: "userSignUp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User email, name and password to register"
        required: true
        schema:
          $ref: "#/definitions/SignUp"
      responses:
        200:
          description: "Successfully Signed Up"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Required fields not satisfied"
        401:
          description: "Wrong Username or Password used"
        409:
          description: "Email entered not available - already registered -"
...
  /auth/store/signin:
    post:
      tags:
      - "Auth"
      summary: "Store Sign In"
      description: "Returns a token for store to be used in further identification"
      operationId: "storeSignIn"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User email and password for sign in"
        required: true
        schema:
          $ref: "#/definitions/SignIn"
      responses:
        200:
          description: "Successfully Logged In"
          type: "object"
          properties:
            detail:
              type: "string" 
            auth:
              type: "boolean"
            token:
              type: "string"
        400:
          description: "Bad request or Required fields not stated"
        401:
          description: "Wrong Username or Password used"
  /auth/store/signup:
    post:
      tags:
      - "Auth"
      summary: "Store Sign Up"
      description: "Creates a new store in the database"
      operationId: "storeSignUp"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Store email, name and password to register"
        required: true
        schema:
          $ref: "#/definitions/SignUp"
      responses:
        200:
          description: "Successfully Signed Up"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Required fields not satisfied"
        401:
          description: "Wrong Username or Password used"
        409:
          description: "Email entered not available - already registered -"
  /auth/reset:
    post:
      tags:
      - "Auth"
      summary: "Resets password"
      description: "Takes a token for reseting password and allows user to enter a new password"
      operationId: "resetPassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Reset Token and new Password"
        required: true
        schema:
          type: "object"
          properties:
            token:
              type: "string"
            token:
              password: "string"
      responses:
        200:
          description: "Password reset successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Required fields not satisfied"
        403:
          description: "Token expired"
  /auth/forget/{model}:
    post:
      tags:
      - "Auth"
      summary: "Forget password"
      description: "Issues a token for user to use in reseting password"
      operationId: "forgetPassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Email to reset password"
        required: true
        schema:
          type: "object"
          properties:
            email:
              type: "string"
              format: "email"
      - name: "model"
        in: "path"
        description: "Whether this is store or user email"
        required: true
        type: "string"
        enum:
        - "user"
        - "store"
      responses:
        200:
          description: "Mail for reset Sent Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Required fields not satisfied"
        404:
          description: "Model or email not found"
...
definitions:
  Success:
    type: "object"
    properties:
      detail:
        type: "string" 
      success:
        type: "boolean"
  SignIn:
    type: "object"
    required:
    - "email"
    - "password"
    properties:
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
        format: "password"
  SignInService:
    type: "object"
    required:
    - "email"
    - "password"
    - "id"
    - "service"
    properties:
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
        format: "password"
      id:
        type: "string"
      service:
        type: "string"
        enum:
          - 'facebook'
          - 'gmail'
  SignUp:
    type: "object"
    required:
    - "email"
    - "name"
    - "password"
    properties:
      email:
        type: "string"
        format: "email"
      name:
        type: "string"
      password:
        type: "string"
        format: "password"
...
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    in: header
    name: "x-access-token"